refactor: get form into separate component
