+++ src/components/Form/Form.module.scss
@@ -1,0 +1,17 @@
+@import '../../styles/variables';
+
+.form {
+  padding: 1rem;
+  display: flex;
+  flex-direction: column;
+  gap: 1.4rem;
+
+  .legend {
+    padding: 0 5px;
+    font-size: 1.1rem;
+  }
+
+  .submit {
+    align-self: center;
+  }
+}
